refactor(domain-health-check): 🔨 split DMARC and SPF verification into partial classes
